Moved worker tasks to outer classes for better readability.
